requirements.txt
Flask==3.0.3
pymongo==4.7.2
python-dotenv
